requirements.txt
# LangChain Core
 langchain
 langchain-core

 langchain-groq
 python-dotenv
 pandas 
 numpy
 scikit-learn

 # Hugging Face Integration
 langchain-huggingface
 transformers
 huggingface-hub

streamlit